Refactor: id 변수 타입 Big Integer로 변경
User 테이블의 PK인 Id를 Big Integer 타입으로 변경

+++ BE/src/main/java/com/codesquad/signup/repository/UserRepository.java
@@ -5,9 +5,10 @@ import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, BigInteger> {
 
     @Query("select user_id from USERS where user_id = :userId")
     Optional<String> findByUserId (@Param("userId") String userId);
